Handle dynamic flags and default values
