+++ main.go
@@ -3,9 +3,9 @@ package main
 import (
 	"fmt"
 	"os"
-	"time"
 
 	"github.com/hatchify/closer"
+	flag "github.com/hatchify/parg"
 	"github.com/hatchify/scribe"
 	"github.com/vroomy/service"
 )
@@ -14,9 +14,12 @@ import (
 const DefaultConfigLocation = "./config.toml"
 
 var (
-	out  *scribe.Scribe
-	svc  *service.Service
+	svc *service.Service
+	cfg *service.Config
+
 	clsr *closer.Closer
+
+	out *scribe.Scribe
 )
 
 func main() {
@@ -24,82 +27,23 @@ func main() {
 	out = scribe.New("Vroomy")
 	out.Notification("Hello there! One moment, initializing..")
 
-	if err = initService(); err != nil {
-		handleError(err)
-	}
-	defer svc.Close()
-
-	clsr = closer.New()
-	go listen()
-	go notifyOfListening()
-
-	if err = clsr.Wait(); err != nil {
-		handleError(err)
-	}
-
-	out.Notification("*Catch*")
-	out.Notification("Close request received, one moment..")
-
-	if err = svc.Close(); err != nil {
-		err = fmt.Errorf("error encountered while closing service: %v", err)
-		handleError(err)
-	}
-
-	out.Success("Service has been closed")
-	os.Exit(0)
-}
-
-func initService() (err error) {
 	configLocation := os.Getenv("VROOMY_CONFIG")
 	if len(configLocation) == 0 {
 		configLocation = DefaultConfigLocation
 	}
 
-	var cfg *service.Config
 	if cfg, err = service.NewConfig(configLocation); err != nil {
 		err = fmt.Errorf("error encountered while reading configuration: %v", err)
 		return
 	}
 
-	out.Notification("Starting service")
-	if svc, err = service.New(cfg); err != nil {
-		err = fmt.Errorf("error encountered while initializing service: %v", err)
-		return
+	var cmd *flag.Command
+	if cmd, err = commandFromArgs(); err != nil {
+		help(cmd)
+		handleError(err)
 	}
 
-	return
-}
-
-func listen() {
-	var err error
-	if err = svc.Listen(); err == nil {
-		return
+	if err = cmd.Exec(); err != nil {
+		handleError(err)
 	}
-
-	err = fmt.Errorf("error encountered while attempting to listen to HTTP: %v", err)
-	clsr.Close(err)
-}
-
-func notifyOfListening() {
-	time.Sleep(time.Millisecond * 300)
-	msg := getListeningMessage(svc.Port(), svc.TLSPort())
-	out.Successf("HTTP is now listening on %s", msg)
-}
-
-func getListeningMessage(port, tlsPort uint16) (msg string) {
-	switch {
-	case port > 0 && tlsPort > 0:
-		msg = fmt.Sprintf("ports %d (HTTP) and %d (HTTPS)", port, tlsPort)
-	case port > 0:
-		msg = fmt.Sprintf("port %d (HTTP)", port)
-	case tlsPort > 0:
-		msg = fmt.Sprintf("port %d (HTTPS)", tlsPort)
-	}
-
-	return
-}
-
-func handleError(err error) {
-	out.Errorf("Fatal error encountered: %v", err)
-	os.Exit(1)
 }
